# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Momentum Terminal API",
    "description": "US equities 12-1 momentum research and INTERNAL paper simulation. No brokerage connectivity.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/paper/init": {
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Paper Init",
        "operationId": "paper_init_api_paper_init_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/paper/advance": {
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Paper Advance",
        "operationId": "paper_advance_api_paper_advance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperAdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/paper/config": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Paper Config",
        "operationId": "paper_config_api_paper_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperConfigResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "portfolio"
        ],
        "summary": "Paper Config Update",
        "operationId": "paper_config_update_api_paper_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvanceResponse": {
        "properties": {
          "processed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Processed"
          },
          "processed_count": {
            "type": "integer",
            "title": "Processed Count"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "stopped_reason": {
            "type": "string",
            "title": "Stopped Reason"
          },
          "stopped_explanation": {
            "type": "string",
            "title": "Stopped Explanation"
          },
          "pending_plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Plan Id"
          }
        },
        "type": "object",
        "required": [
          "processed",
          "processed_count",
          "as_of",
          "stopped_reason",
          "stopped_explanation",
          "pending_plan_id"
        ],
        "title": "AdvanceResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaperAdvanceRequest": {
        "properties": {
          "until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until",
            "description": "Last session to process (default: latest stored data)"
          },
          "max_sessions": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Sessions"
          }
        },
        "type": "object",
        "title": "PaperAdvanceRequest"
      },
      "PaperConfigResponse": {
        "properties": {
          "config_id": {
            "type": "integer",
            "title": "Config Id"
          },
          "config": {
            "$ref": "#/components/schemas/StrategyConfig"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "config_id",
          "config",
          "history"
        ],
        "title": "PaperConfigResponse"
      },
      "PaperInitRequest": {
        "properties": {
          "inception_session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inception Session",
            "description": "Session at whose close cash is funded; default latest data session"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StrategyConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "PaperInitRequest"
      },
...
      "PortfolioMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "initial_capital": {
            "type": "number",
            "title": "Initial Capital"
          },
          "cash": {
            "type": "number",
            "title": "Cash"
          },
          "inception_session": {
            "type": "string",
            "title": "Inception Session"
          },
          "as_of_session": {
            "type": "string",
            "title": "As Of Session"
          },
          "config_id": {
            "type": "integer",
            "title": "Config Id"
          },
          "cum_costs": {
            "type": "number",
            "title": "Cum Costs"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "provider",
          "status",
          "initial_capital",
          "cash",
          "inception_session",
          "as_of_session",
          "config_id",
          "cum_costs",
          "created_at",
          "notes"
        ],
        "title": "PortfolioMeta"
      },
...
      "StrategyConfig": {
        "properties": {
          "initial_capital": {
            "type": "number",
            "maximum": 10000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Initial Capital",
            "description": "Starting virtual cash, USD",
            "default": 100000.0
          },
          "top_n": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Top N",
            "description": "Number of top-ranked eligible stocks to hold",
            "default": 50
          },
          "lookback_sessions": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Lookback Sessions",
            "description": "Older momentum anchor: sessions before signal",
            "default": 252
          },
          "skip_sessions": {
            "type": "integer",
            "maximum": 252.0,
            "minimum": 0.0,
            "title": "Skip Sessions",
            "description": "Recent anchor: sessions before signal (skip month)",
            "default": 21
          },
          "min_history_sessions": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Min History Sessions",
            "description": "Valid bars required before the signal session",
            "default": 252
          },
          "min_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Price",
            "description": "Raw (unadjusted) close at the signal session must exceed this, USD",
            "default": 5.0
          },
          "adv_window": {
            "type": "integer",
            "maximum": 252.0,
            "minimum": 1.0,
            "title": "Adv Window",
            "description": "Sessions in the average dollar-volume window",
            "default": 20
          },
          "min_adv_usd": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Adv Usd",
            "description": "Minimum trailing average daily dollar volume, USD",
            "default": 5000000.0
          },
          "slippage_bps": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Slippage Bps",
            "description": "Assumed adverse slippage vs. the open, basis points",
            "default": 10.0
          },
          "commission_per_order": {
            "type": "number",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Commission Per Order",
            "description": "Assumed flat commission per fill, USD",
            "default": 0.0
          },
          "commission_bps": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Commission Bps",
            "description": "Assumed commission as bps of traded value",
            "default": 0.0
          },
          "min_session_coverage": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Session Coverage",
            "description": "Share of the point-in-time universe that must have a bar at the signal session; otherwise the rebalance is blocked",
            "default": 0.9
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date",
            "description": "Backtest start (ISO date); ignored by the paper ledger"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date",
            "description": "Backtest end (ISO date); ignored by the paper ledger"
          },
          "benchmark_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Symbol",
            "description": "Benchmark; default = provider's SPY (or demo stand-in)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StrategyConfig",
        "description": "All tunable inputs of the 12-1 momentum strategy.\n\nDefaults are the documented v1 specification. Every backtest run and every\npaper-portfolio configuration version stores the full JSON of this model."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}